* Added Basic Health Stats

+++ Packages/monitor-backend/src/gql/_general.rs
@@ -11,7 +11,7 @@ use rust_shared::serde;
 use rust_shared::serde::{Deserialize, Serialize};
 use rust_shared::serde_json::json;
 use rust_shared::tokio_postgres::Client;
-use rust_shared::utils::_k8s::{get_k8s_pod_basic_infos, get_or_create_k8s_secret, try_get_k8s_secret};
+use rust_shared::utils::_k8s::{exec_command_in_another_pod, get_k8s_pod_basic_infos, get_or_create_k8s_secret, try_get_k8s_secret};
 use rust_shared::utils::futures::make_reliable;
 use rust_shared::utils::general_::extensions::ToOwnedV;
 use rust_shared::utils::k8s::cert_handling::get_reqwest_client_with_k8s_certs;
@@ -25,7 +25,7 @@ use std::fs::File;
 use std::io::Read;
 use std::path::Path;
 use std::str::FromStr;
-use std::{env, fs};
+use std::{env, fs, vec};
 use std::{pin::Pin, task::Poll, time::Duration};
 use tracing::{error, info};
 
@@ -49,36 +49,72 @@ pub fn ensure_admin_key_is_correct(admin_key: String, log_message_if_wrong: bool
 	Ok(())
 }
 
-wrap_slow_macros! {
+// wrap_slow_macros! {
 
 // queries
 // ==========
+
+#[derive(SimpleObject, Debug)]
+struct DfRecord {
+	filesystem: String,
+	blocks_1k: u64,
+	used: u64,
+	available: u64,
+	use_percent: u8,
+	mounted_on: String,
+}
+
+impl FromStr for DfRecord {
+	type Err = &'static str;
+
+	fn from_str(line: &str) -> Result<Self, Self::Err> {
+		let parts: Vec<&str> = line.split_whitespace().collect();
+		if parts.len() != 6 {
+			return Err("Line does not have 6 parts");
+		}
+
+		let filesystem = parts[0].to_string();
+		let blocks_1k = parts[1].parse().map_err(|_| "Failed to parse blocks_1k")?;
+		let used = parts[2].parse().map_err(|_| "Failed to parse used")?;
+		let available = parts[3].parse().map_err(|_| "Failed to parse available")?;
+		let use_percent_str = parts[4].trim_end_matches('%');
+		let use_percent = use_percent_str.parse().map_err(|_| "Failed to parse use_percent")?;
+		let mounted_on = parts[5].to_string();
+
+		Ok(DfRecord { filesystem, blocks_1k, used, available, use_percent, mounted_on })
+	}
+}
 
 #[derive(Default)]
 pub struct QueryShard_General;
 #[Object]
 impl QueryShard_General {
 	/// async-graphql requires there to be at least one entry under the Query section
-	async fn empty(&self) -> &str { "" }
+	async fn empty(&self) -> &str {
+		""
+	}
 
 	async fn mtxResults(&self, ctx: &async_graphql::Context<'_>, admin_key: String, start_time: f64, end_time: f64) -> Result<Vec<MtxDataForAGQL>, GQLError> {
 		ensure_admin_key_is_correct(admin_key, true)?;
 
 		let app_state = ctx.data::<AppStateArc>().unwrap();
 		let mtx_results = app_state.mtx_results.read().await.to_vec();
-		let mtx_results_filtered = mtx_results.into_iter().filter(|mtx| {
-			for lifetime in mtx.section_lifetimes.values() {
-				let section_start = lifetime.start_time;
-				let section_end = match lifetime.duration {
-					Some(duration) => section_start + duration,
-					None => f64::MAX, // for this context of filtering, consider a not-yet-ended section to extend to max-time
-				};
-				if section_start < end_time && section_end > start_time {
-					return true;
+		let mtx_results_filtered = mtx_results
+			.into_iter()
+			.filter(|mtx| {
+				for lifetime in mtx.section_lifetimes.values() {
+					let section_start = lifetime.start_time;
+					let section_end = match lifetime.duration {
+						Some(duration) => section_start + duration,
+						None => f64::MAX, // for this context of filtering, consider a not-yet-ended section to extend to max-time
+					};
+					if section_start < end_time && section_end > start_time {
+						return true;
+					}
 				}
-			}
-			false
-		}).collect_vec();
+				false
+			})
+			.collect_vec();
 		let mtx_results_filtered_transformed = mtx_results_filtered.into_iter().map(|mtx| MtxDataForAGQL::from_base(&mtx)).collect_vec();
 		Ok(mtx_results_filtered_transformed)
 	}
@@ -139,6 +175,19 @@ impl QueryShard_General {
 		let log_entries = query_loki(query, startTime, endTime, limit).await?;
 		Ok(QueryLokiResult { logEntries: log_entries })
 	}
+
+	async fn healthStats(&self, _ctx: &async_graphql::Context<'_>, admin_key: String) -> Result<Vec<DfRecord>, GQLError> {
+		ensure_admin_key_is_correct(admin_key, true)?;
+
+		let target_pod =
+			get_k8s_pod_basic_infos("postgres-operator", true).await.context("Failed to retrieve basic-info of the k8s pods.")?.into_iter().find(|a| a.name.starts_with("debate-map-instance1")).map(|a| a.name).ok_or_else(|| anyhow!("Could not find debate-map-instance1-XXX pod."))?;
+		let container = "database";
+		let df_output = exec_command_in_another_pod("postgres-operator", &target_pod, Some(container), "df", vec![], true).await.context("Failed to run pg_dump command in PG pod.")?;
+
+		let df = df_output.lines().skip(1).filter_map(|line| line.parse().ok()).collect();
+
+		Ok(df)
+	}
 }
 
 #[derive(InputObject, Deserialize)]
@@ -147,7 +196,7 @@ pub struct QueryLokiInput {
 	query: String,
 	startTime: i64,
 	endTime: Option<i64>,
-	limit: Option<i64>
+	limit: Option<i64>,
 }
 #[derive(SimpleObject, Debug)]
 struct QueryLokiResult {
@@ -188,11 +237,7 @@ pub async fn query_loki(query: String, startTime: i64, endTime: i64, limit: i64)
 
 pub async fn get_basic_info_from_app_server() -> Result<JSONValue, Error> {
 	let client = new_hyper_client_http();
-	let req = rust_shared::hyper::Request::builder()
-		.method(Method::GET)
-		.uri("http://dm-app-server.default.svc.cluster.local:5110/basic-info")
-		.header("Content-Type", "application/json")
-		.body(json!({}).to_string().into())?;
+	let req = rust_shared::hyper::Request::builder().method(Method::GET).uri("http://dm-app-server.default.svc.cluster.local:5110/basic-info").header("Content-Type", "application/json").body(json!({}).to_string().into())?;
 	let res = client.request(req).await?;
 	let res_as_json_str = body_to_str(res.into_body()).await?;
 	let res_as_json = JSONValue::from_str(&res_as_json_str)?;
@@ -210,15 +255,16 @@ pub async fn tell_k8s_to_restart_app_server() -> Result<JSONValue, Error> {
 	// supply "true", to ignore pods that are already terminating [edit: this doesn't actually work, because terminating pods still show up as "running"; see below for working fix, through use of creation-time field]
 	let k8s_pods = get_k8s_pod_basic_infos("default", true).await?;
 	info!("Got k8s_pods: {:?}", k8s_pods);
-	let app_server_pod_info = k8s_pods.iter().filter(|a| a.name.starts_with("dm-app-server-"))
-		.sorted_by_key(|a| &a.creation_time_str).last() // sort by creation-time, then find last (this way we kill the most recent, if multiple pod matches exist)
-		.ok_or(anyhow!("App-server pod not found in list of active pods."))?.to_owned();
+	let app_server_pod_info = k8s_pods
+		.iter()
+		.filter(|a| a.name.starts_with("dm-app-server-"))
+		.sorted_by_key(|a| &a.creation_time_str)
+		.last() // sort by creation-time, then find last (this way we kill the most recent, if multiple pod matches exist)
+		.ok_or(anyhow!("App-server pod not found in list of active pods."))?
+		.to_owned();
 
 	let client = get_reqwest_client_with_k8s_certs()?;
-	let req = client.delete(format!("https://{k8s_host}:{k8s_port}/api/v1/namespaces/default/pods/{}", app_server_pod_info.name))
-		.header("Content-Type", "application/json")
-		.header("Authorization", format!("Bearer {token}"))
-		.body(json!({}).to_string()).build()?;
+	let req = client.delete(format!("https://{k8s_host}:{k8s_port}/api/v1/namespaces/default/pods/{}", app_server_pod_info.name)).header("Content-Type", "application/json").header("Authorization", format!("Bearer {token}")).body(json!({}).to_string()).build()?;
 	let res = client.execute(req).await?;
 
 	let res_as_json_str = res.text().await?;
@@ -242,8 +288,10 @@ struct StartMigration_Result {
 	migrationID: String,
 }
 
-#[derive(Default)] pub struct MutationShard_General;
-#[Object] impl MutationShard_General {
+#[derive(Default)]
+pub struct MutationShard_General;
+#[Object]
+impl MutationShard_General {
 	/*async fn clearLogEntries(&self, ctx: &async_graphql::Context<'_>, admin_key: String) -> Result<GenericMutation_Result, GQLError> {
 		ensure_admin_key_is_correct(admin_key, true)?;
 
@@ -261,9 +309,7 @@ struct StartMigration_Result {
 
 		tell_k8s_to_restart_app_server().await?;
 
-		Ok(GenericMutation_Result {
-			message: "success".to_string(),
-		})
+		Ok(GenericMutation_Result { message: "success".to_string() })
 	}
 
 	async fn clearMtxResults(&self, ctx: &async_graphql::Context<'_>, admin_key: String) -> Result<GenericMutation_Result, GQLError> {
@@ -273,9 +319,7 @@ struct StartMigration_Result {
 		let mut mtx_results = app_state.mtx_results.write().await;
 		mtx_results.clear();
 
-		Ok(GenericMutation_Result {
-			message: "success".to_string(),
-		})
+		Ok(GenericMutation_Result { message: "success".to_string() })
 	}
 
 	async fn startMigration(&self, ctx: &async_graphql::Context<'_>, admin_key: String, to_version: usize) -> Result<StartMigration_Result, GQLError> {
@@ -291,9 +335,7 @@ struct StartMigration_Result {
 		}
 		let migration_id = migration_result?;
 
-		Ok(StartMigration_Result {
-			migrationID: migration_id,
-		})
+		Ok(StartMigration_Result { migrationID: migration_id })
 	}
 
 	async fn executeTestSequence(&self, ctx: &async_graphql::Context<'_>, admin_key: String, sequence: TestSequence) -> Result<GenericMutation_Result, GQLError> {
@@ -307,9 +349,7 @@ struct StartMigration_Result {
 			error!("Got error while running migration:{}", err);
 		}*/
 
-		Ok(GenericMutation_Result {
-			message: "success".to_owned(),
-		})
+		Ok(GenericMutation_Result { message: "success".to_owned() })
 	}
 }
 
@@ -331,18 +371,17 @@ pub struct TestingLogEntry {
 }
 //#[derive(Clone, SimpleObject)] pub struct GQLSet_LogEntry { pub nodes: Vec<LogEntry> }
 
-#[derive(Default)] pub struct SubscriptionShard_General;
-#[Subscription] impl SubscriptionShard_General {
+#[derive(Default)]
+pub struct SubscriptionShard_General;
+#[Subscription]
+impl SubscriptionShard_General {
 	#[graphql(name = "_ping")]
 	async fn _ping(&self, _ctx: &async_graphql::Context<'_>) -> impl Stream<Item = PingResult> {
 		let pong = "pong".to_owned();
 		// create the listed file in the app-server pod (eg. using Lens), if you've made an update that you need all clients to refresh for
 		let refreshPage = Path::new("./refreshPageForAllUsers_enabled").exists();
 
-		stream::once(async move { PingResult {
-			pong,
-			refreshPage,
-		} })
+		stream::once(async move { PingResult { pong, refreshPage } })
 	}
 
 	async fn logEntries<'a>(&self, ctx: &'a async_graphql::Context<'_>, admin_key: String) -> impl Stream<Item = Result<Vec<LogEntry>, SubError>> + 'a {
@@ -459,4 +498,4 @@ pub struct TestingLogEntry {
 	}
 }
 
-}
+// }
